renamed index.js to main.js in utils
